animation works, open non-animated collapsible doesn't
