+++ apps/website/src/routes/docs/headless/collapsible/snippets/collapsible.css
@@ -35,18 +35,22 @@
 
 .collapsible-content {
   width: 100%;
-  border: 2px dotted hsla(var(--primary) / 1);
   font-weight: 500;
   background: hsla(var(--primary) / 0.2);
   border-radius: calc(var(--border-radius) / 2);
   max-width: var(--select-width);
   color: hsl(var(--foreground));
   overflow: hidden;
+  margin-top: -2px;
+}
+
+.collapsible:has(.collapsible-content:not([hidden])) .collapsible-trigger {
+  border-bottom: 2px dotted;
 }
 
 .collapsible-content-outline {
   padding: 0.5rem;
-  border: 2px dotted hsla(var(--foreground) / 0.6);
+  border: 2px dotted hsla(var(--primary) / 1);
 }
 
 /* animations only */
@@ -57,18 +61,22 @@
 @keyframes collapsible-open {
   0% {
     height: 0;
+    margin-top: 0;
   }
   100% {
     height: var(--qwikui-collapsible-content-height);
+    margin-top: -2px;
   }
 }
 
 @keyframes collapsible-closed {
   0% {
     height: var(--qwikui-collapsible-content-height);
+    margin-top: -2px;
   }
   100% {
     height: 0;
+    margin-top: 0;
   }
 }
