--- packages/kit-headless/src/utils/get-hidden-height.ts
@@ -1,20 +1,0 @@
-/* similar to React's useLayoutEffect, we want to get the height of a hidden element */
-export function getHiddenHeight(el: HTMLElement) {
-  const clone = el.cloneNode(true) as HTMLElement;
-
-  Object.assign(clone.style, {
-    overflow: 'visible',
-    height: 'auto',
-    maxHeight: 'none',
-    opacity: '0',
-    visibility: 'hidden',
-    display: 'block',
-  });
-
-  el.after(clone);
-  const height = clone.offsetHeight;
-
-  clone.remove();
-
-  return height;
-}
